Plan 2023: Domain upsell redone
